infra/sensor_mocker/pyproject.toml
[tool.poetry]
name = "sensor-mocker"
version = "0.1.0"
description = "A basic package for mocking sensor data over the mqtt protocol"
authors = ["darth-veitcher <1722315+darth-veitcher@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
